+++ test/tree-shaking/4-of-1.mjs
@@ -1,0 +1,11 @@
+/* eslint-disable import/extensions */
+/* eslint-disable unicorn/import-index */
+import { DataBitstream } from '../../esm/index.js';
+
+const main = () => {
+  const dbs = new DataBitstream();
+
+  return [dbs];
+};
+
+export default main;